Hide both the occurrance and the bodies folder
This prevents bodies in hidden subcomponents from showing when you select
a top-level component.

The drawback is that its a bit harder to see the hidden subcomponents.
Previously, you could simply select the hidden subcomponent in the browser
and see the outline. Now, you have to unhide the bodies folder, and then
select the subcomponent in the browser.