Tweak the start position of the threads in the Thread component
This makes is so that the "extra" amount of thread at the start and end
of the threads is always a multiple of the pitch (typically 1 pitch).

This helps ensure a more consistent thread positioning for threads of
the same pitch, which can help line up the threads on different parts.