Restructure the code into the fscad.fscad module.
This splits out the add-in script from the actual fscad implementation,
which is now nested under the fscad package.

Additionally, this brings out the test_utils package and also adds it
to the fscad package, for use by external clients of the package.

And finally, the name of the add-in changed from fscad to fscad_addin. This
was needed in order to ensure IntelliJ IDEA/PyCharm resolved the "fscad"
import to the new fscad package, not the top-level fscad.py add-in script.


This change unfortunately breaks existing code, in a few ways.
- any fscad imports will need to be changed to fscad.fscad
- the exported names of the module was cleaned up quite a bit, so that
internal imports aren't "leaked" when doing an import * from fscad. This
may cause errors in existing code that was relying on these leaked import.
e.g. a lot of individual classes from adsk.core and adsk.fusion were being
leaked this way.

Additionally, the fscad add-in will need to be re-imported into fusion,
since the name of the add-in changed from fscad to fscad_addin.